refactor: split message queue tests into modules
