+++ crates/opengoose-persistence/src/message_queue/tests/mod.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+use crate::test_helpers::{ensure_session, test_db};
+use std::sync::Arc;
+
+mod broadcasts;
+mod concurrency;
+mod delegations;
+mod delivery;
+mod queries;
+mod status;
